pyproject.toml
[project]
name = "strava-segments"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9",
    "httpx>=0.27",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
]

[project.scripts]
strava-segments = "strava_segments.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
